==> scripts/20_select_data.Rmd <==
---
title : "MiSTE 'Dynamique des poissons des rivières'"
subtitle: "Sélection des données faunistiques"
date: "`r format(Sys.time(), '%Y-%m-%d')`"
output:
  bookdown::html_document2:
    toc: true
    toc_float: true
#  bookdown::word_document2
always_allow_html: true
params:
  start_year: 1900
  end_year: 2024
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
```

# Setup

## Load packages and functions

```{r}
# if necessary
# devtools::install_github("pascalirz/aspe")
library(aspe)
library(tidyverse)

# install custom functions
# source("../R/functions.R")
```

## Load data

```{r}
# retrieve most recent data file from repo 
rdata_tables <- misc_nom_dernier_fichier(
  repertoire = "../../../raw_data/rdata",
  pattern = "^tables")
# load it
load(rdata_tables)
```

Data file : `r str_replace(rdata_tables, "../../../../projets/ASPE/raw_data/rdata/", "")`


# Spatial selection

# Spatialise sampling points

The dataframe of the selected sampling points is transformed into a spatial object of class `sf`.

```{r}
points_geo <- point_prelevement %>%
  filter(pop_id %in% ids_points) %>%
  left_join(y = ref_type_projection,
            by = c("pop_typ_id" = "typ_id")) %>%
  geo_convertir_coords_df(
    var_x = pop_coordonnees_x,
    var_y = pop_coordonnees_y,
    var_id = pop_id,
    var_crs_initial = typ_code_epsg,
    crs_sortie = 4326
  ) %>%
  rename(x_wgs84 = X,
         y_wgs84 = Y) %>%
  sf::st_as_sf(coords = c("x_wgs84", "y_wgs84"),
               crs = 4326) %>%
  left_join(y = points)
```

```{r}
mapview::mapview(points_geo)
```

```{r}
points_geo <- points_geo %>% 
  sf::st_crop(
    xmin = -5.4,
    xmax = 8.6,
    ymin = 41,
    ymax = 52
  )
```

```{r}
mapview::mapview(points_geo)

ids_points <- unique(points_geo$pop_id)
```


# Assembly of datasets


From the overall dataset of the Aspe database, only 3 monitoring networks are considered here.

-   Réseau de Contrôle de Surveillance (RCS)
-   Réseau Hydrobiologique Piscicole (RHP)
-   Réseau de Référence Pérenne (RRP)

The former two are considered representative of the diversity of the rivers ecosystems in continental France. The RRP (perennial reference network) encompasses only minimally-impacted sites.

Note that one site can contribute to more than one network (true also for one survey).

The data before 2007 were discarded because there was too much turnover between sites and the field protocols were not settled yet.

## Join tables and filter

### Getting started

We start by creating elementary_samples "gateway" dataframe connecting tables ids + additional variables.

This dataframe is then filtered to keep only:

-   the surveys belonging to the representative and reference networks
-   since 2007 for the monitoring networks and since `r params$refnet_start` for the reference network
-   with FBI rating

The selected sampling networks (=aims) are:

```{r}
my_aims <- c("RCS – Réseau de Contrôle de Surveillance",
             "RHP – Réseau Hydrobiologique Piscicole",
             "RRP – Réseau de Référence Pérenne")
```

Assembly and filtering of the dataset.

```{r}
elementary_samples <- mef_creer_passerelle() %>% 
  select(sta_id:pre_id) %>% # remove the unnecessary variables
  distinct() %>% # remove duplicates due to multiple fish batches per survey
  mef_ajouter_ope_date() %>% 
  mef_ajouter_objectif() %>% 
  mef_ajouter_ipr() %>% 
  filter(annee > params$start_year,
         annee <= params$end_year,
         obj_libelle %in% my_aims,
         pop_id %in% ids_points)
```

A few sites have multiple surveys some years. We keep only the last (autumn) one.

```{r}
elementary_samples <- elementary_samples %>% 
  group_by(pop_id, annee, obj_libelle) %>% 
  filter(ope_date == max(ope_date)) %>%   # only last date for sites sampled more than once a year
  ungroup()
```


### Surveys

```{r}
surveys <- elementary_samples %>% 
  select(sta_id:ope_id,
         ope_date,
         annee) %>% 
  filter(pop_id %in% ids_points) %>% 
  distinct()
```

# Data selection

At this stage we kept the sampling points with a minimum of 5 annual surveys.

## Points with enough data

```{r}
points_with_enough_data <- surveys %>% 
  group_by(pop_id) %>% 
  summarise(n_years = n_distinct(annee),
            first_year = min(annee),
            last_year = max(annee)) %>% 
  filter(n_years > 4)

# little checks
table(points_with_enough_data$first_year)
table(points_with_enough_data$last_year)
table(points_with_enough_data$n_years)
```

Ids of the points with enough data

```{r}
ids_points_with_enough_data <- points_with_enough_data %>% 
  pull(pop_id) 
```

## Final data selection

We keep the sampling points either that have been surveyed frequently or that are "reference" sites.


### Surveys

```{r}
surveys <- surveys %>% 
  filter(pop_id %in% ids_points_with_enough_data)
```

### Catches

```{r}
catches <- elementary_samples %>% 
  filter(pop_id %in% ids_points_with_enough_data) %>% 
  mef_ajouter_lots() %>% 
  group_by(sta_id, pop_id, ope_id, annee, esp_code_alternatif) %>% 
    summarise(effectif = sum(lop_effectif), .groups = 'drop')
```


### Points

```{r}
points <- points %>% 
  filter(pop_id %in% ids_points_with_enough_data)

points_geo <- points_geo %>% 
  filter(pop_id %in% ids_points_with_enough_data)
```





# Saving

```{r}
save(catches,
     elementary_samples,
     surveys,
     points,
     points_geo,
     file = "../processed_data/20_selected_data.RData")
```
